data_collect/requirements.txt
requests
pandas
beautifulsoup4
selenium
python-dateutil